% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_transforms.R
\name{calculate_age}
\alias{calculate_age}
\title{Calculate age from birth date}
\usage{
calculate_age(birth_date, ref_date = Sys.Date())
}
\arguments{
\item{birth_date}{Date of birth}

\item{ref_date}{Reference date (default today)}
}
\value{
Integer age in years
}
\description{
Calculate age from birth date
}
