% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/db_queries.R
\name{close_db_connections}
\alias{close_db_connections}
\title{Close database connections}
\usage{
close_db_connections(conns)
}
\arguments{
\item{conns}{List of connection objects from \code{\link{get_db_connections}}}
}
\value{
NULL (invisible)
}
\description{
Safely close CDW and MPI database connections.
}
\examples{
\dontrun{
conns <- get_db_connections()
# ... use connections ...
close_db_connections(conns)
}

}
