% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_transforms.R
\name{create_tooltip}
\alias{create_tooltip}
\title{Create tooltip HTML for an event}
\usage{
create_tooltip(...)
}
\arguments{
\item{...}{Named parameters for tooltip content}
}
\value{
HTML string
}
\description{
Create tooltip HTML for an event
}
