% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/db_queries.R
\name{execute_query}
\alias{execute_query}
\title{Execute a parameterized query with database-appropriate syntax}
\usage{
execute_query(conn, sql, params = list())
}
\arguments{
\item{conn}{Database connection}

\item{sql}{SQL query with ? placeholders}

\item{params}{List of parameters}
}
\value{
Query results as data frame
}
\description{
Execute a parameterized query with database-appropriate syntax
}
