% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filter_helpers.R
\name{filter_by_date_range}
\alias{filter_by_date_range}
\title{Filter events by date range}
\usage{
filter_by_date_range(events, start_date, end_date)
}
\arguments{
\item{events}{Data frame of timeline events}

\item{start_date}{Start of date range}

\item{end_date}{End of date range}
}
\value{
Filtered data frame
}
\description{
Filter events by date range
}
