% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filter_helpers.R
\name{filter_by_dx_pattern}
\alias{filter_by_dx_pattern}
\title{Filter diagnoses by code pattern}
\usage{
filter_by_dx_pattern(events, patient_data, pattern)
}
\arguments{
\item{events}{Data frame of timeline events}

\item{patient_data}{Full patient data list}

\item{pattern}{SQL LIKE pattern (e.g., "E11\%")}
}
\value{
Filtered data frame
}
\description{
Filter diagnoses by code pattern
}
