% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filter_helpers.R
\name{filter_by_encounter_type}
\alias{filter_by_encounter_type}
\title{Filter events by encounter type}
\usage{
filter_by_encounter_type(events, patient_data, enc_types)
}
\arguments{
\item{events}{Data frame of timeline events}

\item{patient_data}{Full patient data list (needed to get encounter info)}

\item{enc_types}{Character vector of encounter types to include}
}
\value{
Filtered data frame
}
\description{
Filter events by encounter type
}
