% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filter_helpers.R
\name{filter_by_event_type}
\alias{filter_by_event_type}
\title{Filter events by event type}
\usage{
filter_by_event_type(events, selected_types)
}
\arguments{
\item{events}{Data frame of timeline events}

\item{selected_types}{Character vector of selected event types}
}
\value{
Filtered data frame
}
\description{
Filter events by event type
}
