% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filter_helpers.R
\name{filter_by_lab_name}
\alias{filter_by_lab_name}
\title{Filter labs by name pattern}
\usage{
filter_by_lab_name(events, patient_data, pattern)
}
\arguments{
\item{events}{Data frame of timeline events}

\item{patient_data}{Full patient data list}

\item{pattern}{Partial match pattern}
}
\value{
Filtered data frame
}
\description{
Filter labs by name pattern
}
