% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filter_helpers.R
\name{filter_by_px_pattern}
\alias{filter_by_px_pattern}
\title{Filter procedures by code pattern}
\usage{
filter_by_px_pattern(events, patient_data, pattern)
}
\arguments{
\item{events}{Data frame of timeline events}

\item{patient_data}{Full patient data list}

\item{pattern}{SQL LIKE pattern}
}
\value{
Filtered data frame
}
\description{
Filter procedures by code pattern
}
