% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filter_helpers.R
\name{filter_by_semantic_results}
\alias{filter_by_semantic_results}
\title{Filter events by semantic filter results}
\usage{
filter_by_semantic_results(events, semantic_results, table_name)
}
\arguments{
\item{events}{Data frame of timeline events}

\item{semantic_results}{Data frame from semantic SQL query}

\item{table_name}{Which table was queried (encounters, diagnoses, etc., or "medications" for UNION)}
}
\value{
Filtered data frame
}
\description{
Filter events by semantic filter results
}
