% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/icd_lookup.R
\name{format_icd_with_description}
\alias{format_icd_with_description}
\title{Format ICD code with description}
\usage{
format_icd_with_description(code, code_type = "10", include_code = TRUE)
}
\arguments{
\item{code}{ICD code}

\item{code_type}{ICD code type}

\item{include_code}{Whether to include the code in output (default TRUE)}
}
\value{
Formatted string
}
\description{
Returns a formatted string with both code and description if available.
}
