% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_transforms.R
\name{format_lab_result}
\alias{format_lab_result}
\title{Format lab result with modifiers and abnormal flags}
\usage{
format_lab_result(
  result_num,
  result_qual,
  result_modifier,
  result_unit,
  abn_ind
)
}
\arguments{
\item{result_num}{Numeric result}

\item{result_qual}{Qualitative result}

\item{result_modifier}{Result modifier (<, >, etc.)}

\item{result_unit}{Result unit}

\item{abn_ind}{Abnormal indicator}
}
\value{
Formatted string
}
\description{
Format lab result with modifiers and abnormal flags
}
