% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_transforms.R
\name{format_vital_content}
\alias{format_vital_content}
\title{Format vital signs for display}
\usage{
format_vital_content(systolic, diastolic, ht, wt, bmi)
}
\arguments{
\item{systolic}{Systolic BP}

\item{diastolic}{Diastolic BP}

\item{ht}{Height}

\item{wt}{Weight}

\item{bmi}{BMI}
}
\value{
Formatted string
}
\description{
Format vital signs for display
}
