% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aggregation.R
\name{get_aggregation_label}
\alias{get_aggregation_label}
\title{Get period label for aggregation}
\usage{
get_aggregation_label(level)
}
\arguments{
\item{level}{Aggregation level}
}
\value{
Human-readable label
}
\description{
Get period label for aggregation
}
