% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filter_helpers.R
\name{get_encounter_types}
\alias{get_encounter_types}
\title{Get unique encounter types from data}
\usage{
get_encounter_types(encounters)
}
\arguments{
\item{encounters}{Encounters data frame}
}
\value{
Character vector of unique encounter types
}
\description{
Get unique encounter types from data
}
