% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/runExample.R
\name{get_sample_data_path}
\alias{get_sample_data_path}
\title{Get Path to Sample Data}
\usage{
get_sample_data_path(file = c("cdw", "mpi"))
}
\arguments{
\item{file}{Which sample data file to return the path for.
Options: "cdw" (PCORnet CDM data) or "mpi" (Master Patient Index).}
}
\value{
Character string with the full path to the sample data file.
}
\description{
Returns the path to the bundled sample DuckDB database files for testing.
}
\examples{
\dontrun{
# Get path to sample CDW database
cdw_path <- get_sample_data_path("cdw")

# Get path to sample MPI database
mpi_path <- get_sample_data_path("mpi")
}

}
