% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/db_queries.R
\name{get_total_event_count}
\alias{get_total_event_count}
\title{Get total event count across all tables}
\usage{
get_total_event_count(patient_data)
}
\arguments{
\item{patient_data}{List of patient data frames}
}
\value{
Integer count of total events
}
\description{
Get total event count across all tables
}
