% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/runExample.R
\name{getViewerURL}
\alias{getViewerURL}
\title{Get Timeline Viewer URL}
\usage{
getViewerURL()
}
\value{
Character URL if running, NULL otherwise.
}
\description{
Get the URL of the currently running background Timeline Viewer.
}
