% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/runExample.R
\name{isViewerRunning}
\alias{isViewerRunning}
\title{Check if Timeline Viewer is Running}
\usage{
isViewerRunning()
}
\value{
Logical TRUE if running, FALSE otherwise.
}
\description{
Check if a background Timeline Viewer is currently running.
}
\examples{
\dontrun{
if (isViewerRunning()) {
  message("Viewer is running at: ", getViewerURL())
}
}

}
