% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/icd_lookup.R
\name{lookup_icd_description}
\alias{lookup_icd_description}
\title{Look up ICD code description}
\usage{
lookup_icd_description(code, code_type = "10")
}
\arguments{
\item{code}{ICD code (character)}

\item{code_type}{ICD code type: "09" for ICD-9, "10" for ICD-10 (from PCORnet DX_TYPE)}
}
\value{
Description string or NA if not found
}
\description{
Looks up the description for an ICD-9 or ICD-10 code using bundled lookup data.
Falls back to the code itself if no description is found.
}
