% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aggregation.R
\name{parse_original_ids}
\alias{parse_original_ids}
\title{Parse original IDs from aggregated event}
\usage{
parse_original_ids(original_ids)
}
\arguments{
\item{original_ids}{Comma-separated string of original event IDs}
}
\value{
Character vector of IDs
}
\description{
Parse original IDs from aggregated event
}
