% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/db_queries.R
\name{qualify_table}
\alias{qualify_table}
\title{Build schema-qualified table name based on database type}
\usage{
qualify_table(table_name, db_type = get_db_type())
}
\arguments{
\item{table_name}{Base table name}

\item{db_type}{Database type ("mssql" or "duckdb")}
}
\value{
Schema-qualified table name
}
\description{
Build schema-qualified table name based on database type
}
