% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/db_queries.R
\name{query_conditions}
\alias{query_conditions}
\title{Query condition data for a patient}
\usage{
query_conditions(conn, patid)
}
\arguments{
\item{conn}{Database connection}

\item{patid}{Patient ID}
}
\value{
Data frame with condition information
}
\description{
Query condition data for a patient
}
