% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/db_queries.R
\name{query_death}
\alias{query_death}
\title{Query death data for a patient}
\usage{
query_death(conn, patid)
}
\arguments{
\item{conn}{Database connection}

\item{patid}{Patient ID}
}
\value{
Data frame with death information (if exists)
}
\description{
Query death data for a patient
}
