% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/db_queries.R
\name{query_death_cause}
\alias{query_death_cause}
\title{Query death cause data for a patient}
\usage{
query_death_cause(conn, patid)
}
\arguments{
\item{conn}{Database connection}

\item{patid}{Patient ID}
}
\value{
Data frame with death cause information (if exists)
}
\description{
Query death cause data for a patient
}
