% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/db_queries.R
\name{query_demographic}
\alias{query_demographic}
\title{Query demographic data for a patient}
\usage{
query_demographic(conn, patid)
}
\arguments{
\item{conn}{Database connection}

\item{patid}{Patient ID}
}
\value{
Data frame with demographic information
}
\description{
Query demographic data for a patient
}
