% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/db_queries.R
\name{query_diagnoses}
\alias{query_diagnoses}
\title{Query diagnosis data for a patient}
\usage{
query_diagnoses(conn, patid)
}
\arguments{
\item{conn}{Database connection}

\item{patid}{Patient ID}
}
\value{
Data frame with diagnosis information
}
\description{
Query diagnosis data for a patient
}
