% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/db_queries.R
\name{query_dispensing}
\alias{query_dispensing}
\title{Query dispensing data for a patient}
\usage{
query_dispensing(conn, patid)
}
\arguments{
\item{conn}{Database connection}

\item{patid}{Patient ID}
}
\value{
Data frame with dispensing information
}
\description{
Query dispensing data for a patient
}
