% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/db_queries.R
\name{query_encounters}
\alias{query_encounters}
\title{Query encounter data for a patient}
\usage{
query_encounters(conn, patid)
}
\arguments{
\item{conn}{Database connection}

\item{patid}{Patient ID}
}
\value{
Data frame with encounter information
}
\description{
Query encounter data for a patient
}
