% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/db_queries.R
\name{query_labs}
\alias{query_labs}
\title{Query lab results for a patient}
\usage{
query_labs(conn, patid)
}
\arguments{
\item{conn}{Database connection}

\item{patid}{Patient ID}
}
\value{
Data frame with lab result information
}
\description{
Query lab results for a patient
}
