% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/db_queries.R
\name{query_prescribing}
\alias{query_prescribing}
\title{Query prescribing data for a patient}
\usage{
query_prescribing(conn, patid)
}
\arguments{
\item{conn}{Database connection}

\item{patid}{Patient ID}
}
\value{
Data frame with prescribing information
}
\description{
Query prescribing data for a patient
}
