% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/db_queries.R
\name{query_procedures}
\alias{query_procedures}
\title{Query procedures data for a patient}
\usage{
query_procedures(conn, patid)
}
\arguments{
\item{conn}{Database connection}

\item{patid}{Patient ID}
}
\value{
Data frame with procedures information
}
\description{
Query procedures data for a patient
}
