% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/db_queries.R
\name{query_source_systems}
\alias{query_source_systems}
\title{Query source system mappings from MPI}
\usage{
query_source_systems(mpi_conn, patid)
}
\arguments{
\item{mpi_conn}{MPI database connection}

\item{patid}{Patient ID (unified PATID = UID in MPI)}
}
\value{
Data frame with source system mappings
}
\description{
Query source system mappings from MPI
}
