% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/db_queries.R
\name{query_vitals}
\alias{query_vitals}
\title{Query vital signs data for a patient}
\usage{
query_vitals(conn, patid)
}
\arguments{
\item{conn}{Database connection}

\item{patid}{Patient ID}
}
\value{
Data frame with vital signs information
}
\description{
Query vital signs data for a patient
}
