% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_transforms.R
\name{safe_coalesce}
\alias{safe_coalesce}
\title{Safe coalesce that handles NULL}
\usage{
safe_coalesce(...)
}
\arguments{
\item{...}{Values to coalesce}
}
\value{
First non-NA value or NA
}
\description{
Safe coalesce that handles NULL
}
