% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_transforms.R
\name{transform_birth_date}
\alias{transform_birth_date}
\title{Create birth date marker for timeline}
\usage{
transform_birth_date(demographic)
}
\arguments{
\item{demographic}{Data frame from query_demographic}
}
\value{
Data frame in timevis format (background type spanning all groups)
}
\description{
Create birth date marker for timeline
}
