% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_transforms.R
\name{transform_conditions}
\alias{transform_conditions}
\title{Transform conditions to timevis format}
\usage{
transform_conditions(conditions)
}
\arguments{
\item{conditions}{Data frame from query_conditions}
}
\value{
Data frame in timevis format
}
\description{
Transform conditions to timevis format
}
