% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_transforms.R
\name{transform_death}
\alias{transform_death}
\title{Create death marker for timeline}
\usage{
transform_death(death)
}
\arguments{
\item{death}{Data frame from query_death}
}
\value{
Data frame in timevis format (background type spanning all groups)
}
\description{
Create death marker for timeline
}
