% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_transforms.R
\name{transform_diagnoses}
\alias{transform_diagnoses}
\title{Transform diagnoses to timevis format}
\usage{
transform_diagnoses(diagnoses)
}
\arguments{
\item{diagnoses}{Data frame from query_diagnoses}
}
\value{
Data frame in timevis format
}
\description{
Transform diagnoses to timevis format
}
