% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_transforms.R
\name{transform_dispensing}
\alias{transform_dispensing}
\title{Transform dispensing to timevis format}
\usage{
transform_dispensing(dispensing)
}
\arguments{
\item{dispensing}{Data frame from query_dispensing}
}
\value{
Data frame in timevis format
}
\description{
Transform dispensing to timevis format
}
