% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_transforms.R
\name{transform_encounters}
\alias{transform_encounters}
\title{Transform encounters to timevis format}
\usage{
transform_encounters(encounters)
}
\arguments{
\item{encounters}{Data frame from query_encounters}
}
\value{
Data frame in timevis format
}
\description{
Transform encounters to timevis format
}
