% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_transforms.R
\name{transform_labs}
\alias{transform_labs}
\title{Transform labs to timevis format}
\usage{
transform_labs(labs)
}
\arguments{
\item{labs}{Data frame from query_labs}
}
\value{
Data frame in timevis format
}
\description{
Transform labs to timevis format
}
