% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_transforms.R
\name{transform_prescribing}
\alias{transform_prescribing}
\title{Transform prescribing to timevis format}
\usage{
transform_prescribing(prescribing)
}
\arguments{
\item{prescribing}{Data frame from query_prescribing}
}
\value{
Data frame in timevis format
}
\description{
Transform prescribing to timevis format
}
