% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_transforms.R
\name{transform_procedures}
\alias{transform_procedures}
\title{Transform procedures to timevis format}
\usage{
transform_procedures(procedures)
}
\arguments{
\item{procedures}{Data frame from query_procedures}
}
\value{
Data frame in timevis format
}
\description{
Transform procedures to timevis format
}
