% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_transforms.R
\name{transform_vitals}
\alias{transform_vitals}
\title{Transform vitals to timevis format}
\usage{
transform_vitals(vitals)
}
\arguments{
\item{vitals}{Data frame from query_vitals}
}
\value{
Data frame in timevis format
}
\description{
Transform vitals to timevis format
}
