% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/icd_lookup.R
\name{icd_data_available}
\alias{icd_data_available}
\title{Check if icd.data package is available}
\usage{
icd_data_available()
}
\value{
Logical TRUE if available
}
\description{
Check if icd.data package is available
}
