% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/db_queries.R
\name{load_patient_data}
\alias{load_patient_data}
\title{Load all patient data}
\usage{
load_patient_data(conns, patid)
}
\arguments{
\item{conns}{Database connections list from \code{\link{get_db_connections}}}

\item{patid}{Patient ID (PATID from the CDM)}
}
\value{
Named list of data frames:
\describe{
\item{demographic}{Patient demographics}
\item{source_systems}{Source system mappings from MPI}
\item{encounters}{Encounter records}
\item{diagnoses}{Diagnosis records}
\item{procedures}{Procedure records}
\item{labs}{Lab result records}
\item{prescribing}{Prescribing records}
\item{dispensing}{Dispensing records}
\item{vitals}{Vital sign records}
\item{conditions}{Condition records}
\item{death}{Death record (if exists)}
\item{death_cause}{Death cause records (if exist)}
}
}
\description{
Load all clinical data for a patient from the PCORnet CDM database.
Returns data from DEMOGRAPHIC, ENCOUNTER, DIAGNOSIS, PROCEDURES,
LAB_RESULT_CM, PRESCRIBING, DISPENSING, VITAL, CONDITION, DEATH,
and DEATH_CAUSE tables.
}
\examples{
\dontrun{
conns <- get_db_connections()
data <- load_patient_data(conns, "PAT0000001")
close_db_connections(conns)
}

}
